requirements.txt
h5py==2.10.0
numpy
pillow
streamlit
tensorflow-cpu==2.8.0
